requirements.txt
pandas>=1.3.0
PyPDF2>=2.0.0
tabula-py>=2.3.0
numpy>=1.20.0